Fixed test coverage.
Moved gitlab and github test from SelfCoreTestCase into
their own test cases.
Added login test in SelfCoreTestCase for a generic platform
